add a `Triple` class to make the code more clear + bugfixes
Previously, the code was relying on 'subject', 'predicate' and
'object' variables. The 'object' variable is not reserved but still a
builtin which might lead to errors. Anyway, without a (data?) class to
hold the triple together it leads to awkward namings or long
names. This also avoid reusing the same variable multiple times which
is a path of hundreds bugs in general.

Mind the fact that the `Triple` instance properties can be variables
`var` and might be serialized depending on the context.

Also, fixed two bugs:

- items were not properly deserialized in some case, see test_seed_subject_lookup

- In the case where the pattern is fully bound and the iterator was not None
  the code assumed the bound items were the parameters