v0.13.0 (#20)
- dev: replace pyenv git clone with submodule

- wiredtiger: check for existing shared object before building

- wiredtiger: run ldconfig

- generic n-tuple store

- tentative fixes of ci...

ref: https://travis-ci.community/t/issue-pipenv-check-fails-due-to-numpy/2120/4

- ... but comment 'pipenv check' because Travis

- README: add image

ref: https://origami.me/jumping-frog/

- make it work with mongodb python3 bindings

That will be soon part of wiredtiger develop branch.
ref: https://jira.mongodb.org/browse/WT-3929

- Travis: wiredtiger use swig to create python3 bindings

Broken.

- rework how to write queries

compose takes as first argument a generator that will yield seed
bindings.

- behold transactional decorator

transactional decorator is inspired from foundationdb. It allows to
compose database queries without having the manage explicitly the
transaction begin and end. That is done by the transactional
decorator. When the first coroutine decorated with transactional is
called, it is responsible to manage the transaction and call
transaction begin. When the coroutine returns, it will call commit.
It will execute the *coroutine* query inside an thread executor, to
try to avoid blocking the main thread. transactional knows that a
transaction is started or not because there is argument that is
initially Hoply object that is wrapped by a Transaction class and
passed downstream, so that further passage through transactional know
that it is not required to start a transaction.

Mind the fact that this transactional decorator is for use in asyncio
context, create your own transactional decorator for other contexts.
Or make a pull-request.

- cosmit :8ball: